# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Game Library API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/library": {
            "get": {
                "tags": [
                    "Library"
                ],
                "summary": "Pobierz listę gier z biblioteki (API)",
                "operationId": "d469bcec0ae40b110c3dc583a00a2e6a",
                "responses": {
                    "200": {
                        "description": "Lista gier w formacie JSON",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "games": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "ID wpisu w bibliotece",
                                                        "type": "integer"
                                                    },
                                                    "game_id": {
                                                        "description": "ID gry (lokalne, z tabeli games)",
                                                        "type": "integer"
                                                    },
                                                    "title": {
                                                        "description": "Tytuł gry (z tabeli games)",
                                                        "type": "string"
                                                    },
                                                    "cover_url": {
                                                        "description": "URL okładki (z tabeli games)",
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "rawg_rating": {
                                                        "description": "Ocena z RAWG (przeliczona na 0-10)",
                                                        "type": "number",
                                                        "format": "float",
                                                        "nullable": true
                                                    },
                                                    "status": {
                                                        "description": "Status gry w bibliotece",
                                                        "type": "string",
                                                        "enum": [
                                                            "to_play",
                                                            "playing",
                                                            "finished"
                                                        ]
                                                    },
                                                    "rating": {
                                                        "description": "Ocena użytkownika (1-10)",
                                                        "type": "integer",
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Library"
                ],
                "summary": "Dodaj grę do biblioteki (API)",
                "operationId": "baaa832b661f395a3380a5b8d1413b49",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rawg_game_id",
                                    "title",
                                    "status"
                                ],
                                "properties": {
                                    "rawg_game_id": {
                                        "description": "ID gry z RAWG API (Battlefield 4)",
                                        "type": "integer",
                                        "example": 3842
                                    },
                                    "title": {
                                        "description": "Tytuł gry",
                                        "type": "string",
                                        "example": "Battlefield 4"
                                    },
                                    "cover_url": {
                                        "description": "URL okładki",
                                        "type": "string",
                                        "example": "https://media.rawg.io/media/games/ac7/ac7b8327343da12c971cfc418f390a11.jpg",
                                        "nullable": true
                                    },
                                    "status": {
                                        "description": "Status gry",
                                        "type": "string",
                                        "example": "playing",
                                        "enum": [
                                            "to_play",
                                            "playing",
                                            "finished"
                                        ]
                                    },
                                    "rating": {
                                        "description": "Ocena użytkownika (1-10)",
                                        "type": "integer",
                                        "example": 8,
                                        "nullable": true,
                                        "maximum": 10,
                                        "minimum": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Gra dodana",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "game": {
                                            "properties": {
                                                "id": {
                                                    "description": "ID wpisu w bibliotece",
                                                    "type": "integer"
                                                },
                                                "game_id": {
                                                    "description": "ID gry (lokalne)",
                                                    "type": "integer"
                                                },
                                                "title": {
                                                    "description": "Tytuł gry",
                                                    "type": "string"
                                                },
                                                "cover_url": {
                                                    "description": "URL okładki",
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "rawg_rating": {
                                                    "description": "Ocena RAWG (przeliczona)",
                                                    "type": "number",
                                                    "nullable": true
                                                },
                                                "status": {
                                                    "description": "Status gry",
                                                    "type": "string",
                                                    "enum": [
                                                        "to_play",
                                                        "playing",
                                                        "finished"
                                                    ]
                                                },
                                                "rating": {
                                                    "description": "Ocena użytkownika",
                                                    "type": "integer",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Błąd walidacji"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}